+++ src/llmtoolkit/__init__.py
@@ -1,3 +1,4 @@
-"""
-LLM-Toolkit main import module
-"""
+from .conversation import Conversation
+from .llm import MistralaiLLM, OpenAILLM
+
+__all__ = ["Conversation", "OpenAILLM", "MistralaiLLM"]