+++ src/llmtoolkit/chain/prompts/prompt.py
@@ -1,25 +1,20 @@
 from collections.abc import Generator
 
-from llmtoolkit.chain.base import Chain
+from llmtoolkit.core import Chain
 from llmtoolkit.core.models import (
     ChainResponse,
     ConversationHistory,
-    ConversationMessage,
-    Roles,
 )
 
 
 class PromptChain(Chain):
     prompt: str
 
-    def _prepare(self, conversation_history: ConversationHistory | None) -> ConversationHistory:
-        conversation_history = conversation_history or ConversationHistory()
-        conversation_history.append(ConversationMessage(role=Roles.ASSISTANT, content=self.prompt))
+    def _prepare(self, conversation_history: ConversationHistory) -> None:
+        conversation_history.add_assistant_message(content=self.prompt)
         return conversation_history
 
-    def generate(
-        self, conversation_history: ConversationHistory | None = None, **kwargs
-    ) -> ChainResponse:
+    def generate(self, conversation_history: ConversationHistory, **kwargs) -> ChainResponse:
         conversation_history = self._prepare(conversation_history)
         return self.chain.generate(conversation_history, **kwargs)