+++ src/llmtoolkit/core/llm.py
@@ -1,7 +1,9 @@
 from abc import ABC
 
-from llmtoolkit.chain import Chain
+from .chain import Chain
 
 
 class BaseLLM(Chain, ABC):
+    api_key: str = "-"
+    host: str | None = None
     model_name: str = "default"