+++ src/llmtoolkit/llm/mistralai.py
@@ -1,103 +1,135 @@
-from collections.abc import Generator
+from collections.abc import AsyncGenerator, Generator
 from typing import Any
 
 import httpx
+from mistralai import UNSET as MISTRAL_UNSET
 from mistralai import Mistral as MistralClient
-from mistralai import ToolCall as MistralToolCall
-from pydantic import PrivateAttr, SecretStr
+from pydantic import PrivateAttr
 
-from llmtoolkit.core.models import ChainResponse, ConversationHistory, ToolCall, ToolCallFunction
+from llmtoolkit.core.llm import BaseLLM
+from llmtoolkit.core.models import (
+    ChainResponse,
+    ConversationHistory,
+)
 from llmtoolkit.exc import StreamReadError
-from llmtoolkit.llm.base import BaseLLM
 
 
 class MistralaiLLM(BaseLLM):
-    api_key: SecretStr
-
     _client: MistralClient = PrivateAttr()
 
     def model_post_init(self, __context: Any) -> None:
-        self._client = MistralClient(api_key=self.api_key.get_secret_value())
+        self._client = MistralClient(api_key=self.api_key)
 
-    @staticmethod
-    def _prepare_tool_calls(tool_calls: list[MistralToolCall] | None) -> list[ToolCall] | None:
-        if not tool_calls:
-            return None
-        return [
-            ToolCall(
-                id=tool_call.id,
-                function=ToolCallFunction(
-                    name=tool_call.function.name,
-                    arguments=tool_call.function.arguments,
-                ),
-            )
-            for tool_call in tool_calls
-        ]
-
-    def generate(self, conversation_history: ConversationHistory | None = None, **kwargs) -> str:
+    def generate(
+        self,
+        conversation_history: ConversationHistory | None = None,
+        *,
+        temperature: float = MISTRAL_UNSET,
+        top_p: float = 1,
+        frequency_penalty: float = 0,
+        presence_penalty: float = 0,
+        max_completion_tokens: int | None = MISTRAL_UNSET,
+        stop: list[str] | None = None,
+        **kwargs: Any,
+    ) -> ChainResponse:
         conversation_history = conversation_history or ConversationHistory()
         response = self._client.chat.complete(
-            model=self.model_name, messages=conversation_history.dump(), **kwargs
+            model=self.model_name,
+            messages=conversation_history.dump(),
+            temperature=temperature,
+            top_p=top_p,
+            frequency_penalty=frequency_penalty,
+            presence_penalty=presence_penalty,
+            max_tokens=max_completion_tokens,
+            stop=stop,
+            **kwargs,
         )
-        return ChainResponse(
-            content=response.choices[0].message.content,
-            metadata={
-                "__tool_calls": self._prepare_tool_calls(response.choices[0].message.tool_calls)
-            },
-        )
+        message = response.choices[0].message
+        return ChainResponse(content=message.content or "")
 
     async def async_generate(
-        self, conversation_history: ConversationHistory | None = None, **kwargs
-    ) -> str:
+        self,
+        conversation_history: ConversationHistory | None = None,
+        *,
+        temperature: float = MISTRAL_UNSET,
+        top_p: float = 1,
+        frequency_penalty: float = 0,
+        presence_penalty: float = 0,
+        max_completion_tokens: int | None = MISTRAL_UNSET,
+        stop: list[str] | None = None,
+        **kwargs: Any,
+    ) -> ChainResponse:
         conversation_history = conversation_history or ConversationHistory()
         response = await self._client.chat.complete_async(
-            model=self.model_name, messages=conversation_history.dump(), **kwargs
+            model=self.model_name,
+            messages=conversation_history.dump(),
+            temperature=temperature,
+            top_p=top_p,
+            frequency_penalty=frequency_penalty,
+            presence_penalty=presence_penalty,
+            max_tokens=max_completion_tokens,
+            stop=stop,
+            **kwargs,
         )
-        return ChainResponse(
-            content=response.choices[0].message.content,
-            metadata={
-                "__tool_calls": self._prepare_tool_calls(response.choices[0].message.tool_calls)
-            },
-        )
+        message = response.choices[0].message
+        return ChainResponse(content=message.content or "")
 
     def generate_stream(
-        self, conversation_history: ConversationHistory | None = None, **kwargs
-    ) -> Generator[str, None, None]:
+        self,
+        conversation_history: ConversationHistory | None = None,
+        *,
+        temperature: float = MISTRAL_UNSET,
+        top_p: float = 1,
+        frequency_penalty: float = 0,
+        presence_penalty: float = 0,
+        max_completion_tokens: int | None = MISTRAL_UNSET,
+        stop: list[str] | None = None,
+        **kwargs: Any,
+    ) -> Generator[ChainResponse, None, None]:
         conversation_history = conversation_history or ConversationHistory()
         try:
             for chunk in self._client.chat.stream(
                 model=self.model_name,
                 messages=conversation_history.dump(),
+                temperature=temperature,
+                top_p=top_p,
+                frequency_penalty=frequency_penalty,
+                presence_penalty=presence_penalty,
+                max_tokens=max_completion_tokens,
+                stop=stop,
                 **kwargs,
             ):
-                yield ChainResponse(
-                    content=chunk.data.choices[0].delta.content or "",
-                    metadata={
-                        "__tool_calls": self._prepare_tool_calls(
-                            chunk.data.choices[0].delta.tool_calls
-                        )
-                    },
-                )
+                message = chunk.data.choices[0].delta
+                yield ChainResponse(content=message.content or "")
         except httpx.ResponseNotRead:
             raise StreamReadError
 
     async def async_generate_stream(
-        self, conversation_history: ConversationHistory | None = None, **kwargs
-    ) -> Generator[str, None, None]:
+        self,
+        conversation_history: ConversationHistory | None = None,
+        *,
+        temperature: float = MISTRAL_UNSET,
+        top_p: float = 1,
+        frequency_penalty: float = 0,
+        presence_penalty: float = 0,
+        max_completion_tokens: int | None = MISTRAL_UNSET,
+        stop: list[str] | None = None,
+        **kwargs: Any,
+    ) -> AsyncGenerator[ChainResponse, None]:
         conversation_history = conversation_history or ConversationHistory()
         try:
             async for chunk in await self._client.chat.stream_async(
                 model=self.model_name,
                 messages=conversation_history.dump(),
+                temperature=temperature,
+                top_p=top_p,
+                frequency_penalty=frequency_penalty,
+                presence_penalty=presence_penalty,
+                max_tokens=max_completion_tokens,
+                stop=stop,
                 **kwargs,
             ):
-                yield ChainResponse(
-                    content=chunk.data.choices[0].delta.content or "",
-                    metadata={
-                        "__tool_calls": self._prepare_tool_calls(
-                            chunk.data.choices[0].delta.tool_calls
-                        )
-                    },
-                )
+                message = chunk.data.choices[0].delta
+                yield ChainResponse(content=message.content or "")
         except httpx.ResponseNotRead:
             raise StreamReadError